+++ src/FreeElephants/TSxUnit/Mock/AbstractGuiStubBuilder.ts
@@ -1,0 +1,8 @@
+namespace FreeElephants.TSxUnit.Mock {
+
+    export abstract class AbstractGuiStubBuilder implements BuilderInterface {
+        public abstract getMock();
+
+        // abstract
+    }
+}