+++ src/FreeElephants/TSxUnit/TestCase.ts
@@ -1,5 +1,5 @@
 ///<reference path="Assert/Assert.ts"/>
-///<reference path="Mock/Builder.ts"/>
+///<reference path="Mock/ObjectBuilder.ts"/>
 
 /**
  * @author samizdam
@@ -7,16 +7,16 @@
 namespace FreeElephants.TSxUnit {
 
     import Assert = FreeElephants.TSxUnit.Assert.Assert;
-    import Builder = FreeElephants.TSxUnit.Mock.Builder;
+    import Builder = FreeElephants.TSxUnit.Mock.ObjectBuilder;
 
     export abstract class TestCase {
 
         /**
          * Exploit some JS magic: for use default value of assertion message if msg argument not present.
          */
-        private _undefined;
+        protected _undefined;
 
-        private numberOfAsserts: number = 0;
+        protected numberOfAsserts: number = 0;
 
         protected assertTrue(expr: boolean, msg: string = this._undefined): void {
             this.numberOfAsserts++;