+++ src/reference.int.ts
@@ -1,3 +1,3 @@
-///<reference path="../typings/main.d.ts"/>
+///<reference path="../typings/globals/node/index.d.ts"/>
 ///<reference path="reference.ext.ts"/>